apps-py/shipping-receiver-py/requirements-ot.txt
Flask==3.0.0
gunicorn==21.2.0
kafka-python==2.0.2
requests==2.31.0
Werkzeug==3.0.1
opentelemetry-distro==0.44b0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation-flask==0.44b0
opentelemetry-instrumentation-kafka-python==0.44b0
opentelemetry-instrumentation-requests==0.44b0
